% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNoiseAtEventTimes.R
\name{addNoiseAtEventTimes}
\alias{addNoiseAtEventTimes}
\title{Tie removal}
\usage{
addNoiseAtEventTimes(fr, id, from, to)
}
\arguments{
\item{fr}{Data on long format}

\item{id}{Name of column in \code{dataFr} that identifies individuals}

\item{from}{Name of the variable that contains the name of start state for
each at-risk interval}

\item{to}{Name of column with stop time of the at risk period}
}
\value{
Longitudinally refined \code{\link{data.table}} of the input \code{\link{data.table}} \code{fr} with random noise added for tiebreaking.
}
\description{
Removes ties from by adding noise at tied times
}
\examples{
library(data.table)
data(fr1)
fr1 <- as.data.table(fr1)
fr1 <- addNoiseAtEventTimes(fr1)
head(fr1)
}
\author{
Pål Christie Ryalen <p.c.ryalen@medisin.uio.no>
}
