\name{ahw-package}
\alias{ahw-package}
\alias{ahw}
\docType{package}
\title{
\packageTitle{ahw}
}
\description{
\packageDescription{ahw}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ahw}
\packageIndices{ahw}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ahw}

Maintainer: \packageMaintainer{ahw}
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
