% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahw-package.R
\docType{package}
\name{ahw}
\alias{ahw}
\title{Re-weighting Point Processes Based on Additive Hazard Models}
\description{
Estimates continuous time weights for performing causal survival
analysis. For instance, weighted Nelson-Aalen or Kaplan-Meier
estimates can be given a causal interpretation.
}
