% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fFrame}
\alias{fFrame}
\title{Simulated survival data}
\format{A data frame with 300 rows and 6 variables:
\describe{
  \item{id}{Individual}
  \item{from}{Start time}
  \item{to}{Event time}
  \item{from.state}{Start state}
  \item{to.state}{Event state}
  \item{u}{Binary vaiable}
}}
\usage{
fFrame
}
\description{
Individuals whos censoring/death hazards depend on a binary baseline variable u
}
\keyword{datasets}
