% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naReplace.R
\name{naReplace}
\alias{naReplace}
\title{Replaces NA-values in \code{vec} with last non-NA value}
\usage{
naReplace(vec)
}
\arguments{
\item{vec}{Vector of any type}
}
\value{
  \code{numeric} \code{vector} with each \code{NA} entry replaced with the last previous non-\code{NA} entry.
}
\description{
Assumes first element is non-NA
}
\note{
Can be replaced by \code{link[zoo]{na.locf0}}
}
\examples{
naReplace(c(1, 2, 3, NA, NA, 4))
naReplace(c("text", NA, NA))
}
\author{
Pål Christie Ryalen <p.c.ryalen@medisin.uio.no>
}
