% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContWeights.R
\name{plotContWeights}
\alias{plotContWeights}
\title{Plots mean and individual weight trajectories}
\usage{
plotContWeights(
  fr,
  stopTimeName = "to",
  startStatusName = "from.state",
  endStatusName = "to.state",
  idName = "id"
)
}
\arguments{
\item{fr}{Data with weight column}

\item{stopTimeName}{Name of column with stop time of the at risk period}

\item{startStatusName}{Name of the variable that contains the name of start
state for each at-risk interval}

\item{endStatusName}{Name of the variable that contains the name of end state
for each at-risk interval}

\item{idName}{Name of column in \code{dataFr} that identifies individuals}
}

\value{
  No return value.
}

\description{
Plotting all the individual weight trajectories in fr,
along with the mean. Plots weights for assessment.
}
\author{
Pål Christie Ryalen <p.c.ryalen@medisin.uio.no>
}
