% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineTable.R
\name{refineTable}
\alias{refineTable}
\title{Expands data.table}
\usage{
refineTable(dataFr, atRiskState, eventState)
}
\arguments{
\item{dataFr}{data.frame or data.table on long format}

\item{atRiskState}{At risk state(s)}

\item{eventState}{Observed event times}
}
\value{
data.table
}
\description{
Refines dataFr so that each individual at risk get a row for
each of the provided event times
}
\author{
Pål Christie Ryalen <p.c.ryalen@medisin.uio.no>
}
